+++ lib/query.dart
@@ -3,7 +3,7 @@ import 'src/grammar.dart';
 
 export 'src/ast.dart';
 
-final _parser = QueryParser();
+final _parser = QueryGrammarDefinition().build();
 
 /// Parses [input] and returns a parsed [Query].
 Query parseQuery(String input) {